+++ AGENTS.md
@@ -44,7 +44,7 @@ Both files are template v0.4.0 adoptions: entries are **retention-capped** (`Act
 - `MainTabView` (template-skeleton merge file) owns the `NavigationStack`, the Home / Activity / Settings tabs, the brand navigation title, the `mainToolbar` extension (Edit / Spin-to-Win / Add Player, gated on the Home tab), and the add-player/spinner sheets. Per-tab `navigationTitle`/`toolbar` preferences do NOT propagate through a `TabView` to the enclosing stack — put per-tab bar items in `mainToolbar`, not in tab content.
 - `HomeView` (merge-pattern body) is the player roster: lists/edits players (delete, reorder) and shows live balances; presents `PlayerView` pushes. **Edit mode** (`@Binding editMode`) is owned by `MainTabView` (its `mainToolbar` Edit/Done button) but injected with `.environment(\.editMode, …)` at the **List level here** — injecting it on the `TabView` does not activate a tab-hosted List (#30). Swipe-to-delete is armed only in Edit mode, and a player who has exchanged money (`GameSession.hasExchangedMoney`) is locked from individual deletion (roster deletes go through `GameSession`, which keeps the transaction log and appends an Activity Log marker; **Delete All Players** in Settings additionally clears the transaction log for a fresh start).
 - `PlayerView` is the per-player banking screen (collect salary, add/subtract, send to another player).
-- `ViewSupport/` holds presentation config aligned with the template: `AppConfig` (merge file — brand, `dynamicSizeMax`, colors), `CustomLabeledContentStyle`, `ErrorAlertViewModifier`, plus app-owned `PlayerThumbnailView`/`CameraImagePicker`/`KeyboardActionBar` (the shared bar above the keyboard, #35/#37/#42 — carries the focused money field's action, one-tap Add/Subtract/Send with Cancel, or plain Done; a `safeAreaInset(edge: .bottom)` bar, deliberately NOT a `.toolbar(placement: .keyboard)` accessory — the v1 accessory and its in-cell placement rules were retired in #42 after device testing hit the known-broken `placement: .keyboard` plumbing (see that file's history note); screens attach `.keyboardActionBar(focus:action:)` on their Form/container, and shared child sections publish `KeyboardActionBarPreference` rendered by the host's `.keyboardActionBarHost()`) and `IntegerFormatter` (the one shared money-entry `NumberFormatter`, #37).
+- `ViewSupport/` holds presentation config aligned with the template: `AppConfig` (merge file — brand, `dynamicSizeMax`, colors), `CustomLabeledContentStyle`, `ErrorAlertViewModifier`, plus app-owned `PlayerThumbnailView`/`CameraImagePicker`/`KeyboardActionBar` (the shared bar above the keyboard, #35/#37/#42 — carries the focused money field's action, one-tap Add/Subtract/Send with Cancel, or plain Done; a `safeAreaInset(edge: .bottom)` bar, deliberately NOT a `.toolbar(placement: .keyboard)` accessory — the v1 accessory and its in-cell placement rules were retired in #42 after device testing hit the known-broken `placement: .keyboard` plumbing (see that file's history note); screens attach `.keyboardActionBar(focus:action:)` on their Form/container, and shared child sections publish `KeyboardActionBarPreference` rendered by the host's `.keyboardActionBarHost()`) and `MoneyFormatter` (the one shared money-entry `NumberFormatter`, #37 — since #45 it renders committed values grouped with a `$` prefix, matching the money displays, and parses leniently).
 - `Utilities/` — template `DebugLogging` (the `DebugPrintable` protocol's `debugprint(_:)`, release no-op, plus `deviceLog`), template `PreviewConfig` (`isPreview`), and app-owned `SoundPlayer`/`PlayerImageMaker`.
